+++ test/parallel/test_tensorflow2_keras.py
@@ -15,10 +15,8 @@
 
 """Tests for horovod.tensorflow.keras in TensorFlow 2."""
 
-import math
 import tensorflow as tf
 import numpy as np
-import os
 import warnings
 
 from distutils.version import LooseVersion