+++ test/parallel/test_xla_process_sets.py
@@ -28,8 +28,6 @@ from distutils.version import LooseVersion
 import itertools
 import numpy as np
 import os
-import platform
-import math
 import pytest
 import sys
 
@@ -40,70 +38,78 @@ import sys
 os.environ["HOROVOD_ENABLE_XLA_OPS"] = "1"
 
 import tensorflow as tf
-from horovod.tensorflow.util import _executing_eagerly
-
-sys.path.append(os.path.join(os.path.dirname(__file__), os.pardir, 'utils'))
 
 import horovod.tensorflow as hvd
 
 from base_test_tensorflow import *
 
-_IS_TF26 = LooseVersion(tf.__version__) >= LooseVersion('2.6.0')
+from horovod.runner.common.util.env import get_env_rank_and_size
 
-# Set environment variable to enable adding/removing process sets after
-# initializing Horovod.
-os.environ["HOROVOD_DYNAMIC_PROCESS_SETS"] = "1"
+_IS_TF26 = LooseVersion(tf.__version__) >= LooseVersion('2.6.0')
 
 
 @pytest.mark.skipif(not _IS_TF26, reason='TF2.6+ is required')
 class XLAProcessSetsTests(BaseTensorFlowTests):
     """
-    Tests for ops in horovod.tensorflow that add/remove process sets after initialization.
+    Tests for ops in horovod.tensorflow using multiple process sets.
     """
 
     def __init__(self, *args, **kwargs):
         super(XLAProcessSetsTests, self).__init__(*args, **kwargs)
 
+    @classmethod
+    def setUpClass(cls):
+        """Initializes Horovod with two process sets"""
+        _, size = get_env_rank_and_size()
+
+        cls.even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
+        cls.odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
+        cls.even_set = hvd.ProcessSet(cls.even_ranks)
+        cls.odd_set = hvd.ProcessSet(cls.odd_ranks)
+
+        hvd.init(process_sets=[cls.even_set, cls.odd_set])
+
+    def tearDown(self):
+        """Prevent that one process shuts down Horovod too early"""
+        with tf.device("/cpu:0"):
+            b = hvd.allreduce(tf.constant([0.]), name="global_barrier_after_test")
+            _ = self.evaluate(b)
+
     def test_horovod_allreduce_gpu_process_sets(self):
         """ Test on XLA/GPU that allreduce correctly sums if restricted to non-global process sets"""
         # Only do this test if there are GPUs available.
         if not tf.test.is_gpu_available(cuda_only=True):
-            self.skipTest(("No GPUs available"))
+            self.skipTest("No GPUs available")
 
         if int(os.environ.get('HOROVOD_MIXED_INSTALL', 0)):
             # Skip if compiled with CUDA but without HOROVOD_GPU_OPERATIONS.
             self.skipTest("Not compiled with HOROVOD_GPU_OPERATIONS")
 
-        hvd.init()
         local_rank = hvd.local_rank()
         rank = hvd.rank()
         size = hvd.size()
 
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
         def allreduce_gpu_process_set(self, dtype, dim):
-            even_rank_tensor = self.random_uniform(
-                [17] * dim, -100, 100, dtype=dtype)
-            odd_rank_tensor = self.random_uniform(
-                [17] * dim, -100, 100, dtype=dtype)
-            if rank in even_ranks:
+            even_rank_tensor = self.random_uniform([17] * dim, -100, 100)
+            even_rank_tensor = tf.cast(even_rank_tensor, dtype=dtype)
+            odd_rank_tensor = self.random_uniform([17] * dim, -100, 100)
+            odd_rank_tensor = tf.cast(odd_rank_tensor, dtype=dtype)
+            if rank in self.even_ranks:
                 summed = hvd.allreduce(
                     even_rank_tensor,
                     average=False,
-                    process_set=even_set)
-                multiplied = even_rank_tensor * len(even_ranks)
-            if rank in odd_ranks:
+                    process_set=self.even_set)
+                multiplied = even_rank_tensor * len(self.even_ranks)
+            if rank in self.odd_ranks:
                 summed = hvd.allreduce(
-                    odd_rank_tensor, average=False, process_set=odd_set)
-                multiplied = odd_rank_tensor * len(odd_ranks)
-            max_difference = tf.reduce_max(tf.abs(summed - multiplied))
+                    odd_rank_tensor, average=False, process_set=self.odd_set)
+                multiplied = odd_rank_tensor * len(self.odd_ranks)
+            difference = summed - multiplied
+            difference = tf.cast(difference, tf.int32) if dtype == tf.uint8 else difference
+            max_difference = tf.reduce_max(tf.abs(difference))
             return max_difference
 
-        dtypes = [tf.int32, tf.int64, tf.float16, tf.float32, tf.float64]
+        dtypes = [tf.uint8, tf.int8, tf.int32, tf.int64, tf.float16, tf.float32, tf.float64]
         dims = [1, 2, 3]
         for dtype, dim in itertools.product(dtypes, dims):
             with tf.device("/gpu:%d" % local_rank):
@@ -112,8 +118,8 @@ class XLAProcessSetsTests(BaseTensorFlowTests):
 
             # Threshold for floating point equality depends on number of
             # ranks, since we're comparing against precise multiplication.
-            max_process_set_size = max(len(even_ranks), len(odd_ranks))
-            if max_process_set_size <= 3 or dtype in [tf.int32, tf.int64]:
+            max_process_set_size = max(len(self.even_ranks), len(self.odd_ranks))
+            if max_process_set_size <= 3 or dtype in [tf.uint8, tf.int8, tf.int32, tf.int64]:
                 threshold = 0
             elif max_process_set_size < 10:
                 threshold = 1e-4
@@ -126,6 +132,3 @@ class XLAProcessSetsTests(BaseTensorFlowTests):
             diff = self.evaluate(max_difference)
             self.assertTrue(diff <= threshold,
                             "hvd.allreduce produces incorrect results")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)